PERF-4204 PPaaS - Split Test Status page into separate from index at /test/[testId]
- Create controller/pages/test/[testId].tsx as a dedicated test status page
  using the testId as a path param instead of a query param
- Slim controller/pages/index.tsx down to history-only (running/recent/searched
  tests + search form); adds a 307 redirect for legacy /?testId= links
- Rename PAGE_TEST_HISTORY_FORMAT -> PAGE_TEST_STATUS_FORMAT (URL changes from
  /?testId=<id> to /test/<id>); update all 6 callers
- Split wdio acceptance tests: test status/results dropdown tests move from
  index.spec.ts to new teststatus.spec.ts; fix &results=0 -> ?results=0 in URLs
