+++ controller/components/StartTestForm/index.tsx
@@ -6,7 +6,7 @@ import {
   AuthPermissions,
   EnvironmentVariablesFile,
   PAGE_CALENDAR,
-  PAGE_TEST_HISTORY_FORMAT,
+  PAGE_TEST_STATUS_FORMAT,
   PreviousTestData,
   TestData
 } from "../../types";
@@ -505,7 +505,7 @@ export const StartTestForm = ({
         const calendarUrl = PAGE_CALENDAR + "?defaultDate=" + responseData.startTime;
         await router.push(calendarUrl, formatPageHref(calendarUrl));
       } else {
-        const statusUrl = PAGE_TEST_HISTORY_FORMAT(responseData.testId);
+        const statusUrl = PAGE_TEST_STATUS_FORMAT(responseData.testId);
         await router.push(statusUrl, formatPageHref(statusUrl));
       }
     } catch (error) {
@@ -701,7 +701,7 @@ export const StartTestForm = ({
             lastUpdated: previousTestData?.lastUpdated
           } } />
           <LinkButton theme={{ buttonFontSize: "1.25rem", buttonWidth: "200px", buttonHeight: "50px" , buttonMargin: "10px"}}
-                      href={PAGE_TEST_HISTORY_FORMAT(state.testId)}>Test Status</LinkButton>
+                      href={PAGE_TEST_STATUS_FORMAT(state.testId)}>Test Status</LinkButton>
         </Column>}
       </Row>
       {previousTestData && <Row data-testid="prior-yaml-section">
