+++ controller/src/testscheduler.ts
@@ -3,7 +3,7 @@ import {
   AuthPermissions,
   ErrorResponse,
   PAGE_START_TEST_FORMAT,
-  PAGE_TEST_HISTORY_FORMAT,
+  PAGE_TEST_STATUS_FORMAT,
   ScheduledTestData,
   ScheduledTestRecurrence,
   StoredTestData,
@@ -78,7 +78,7 @@ export function getTestsToRun (scheduledTests: Map<string, TestSchedulerItem> | 
   return Array.from(scheduledTests.values()).filter((testItem: TestSchedulerItem) => testItem.nextStart <= Date.now());
 }
 
-const createHistoricalUrl = (testId: string) => PAGE_TEST_HISTORY_FORMAT(testId);
+const createHistoricalUrl = (testId: string) => PAGE_TEST_STATUS_FORMAT(testId);
 const createEventUrl = (testId: string) => PAGE_START_TEST_FORMAT(testId, true);
 const createHistoricalEvent = (testId: string, yamlFile: string, startTime: number, endTime: number, testStatus: TestStatus): EventInput => ({
   /** String or Integer. Will uniquely identify your event. Useful for getEventById. */
